+++ packages/smarthr-ui/src/components/Button/client/index.ts
@@ -1,0 +1,1 @@
+export { ButtonWrapper } from './ButtonWrapper'
